+++ lib/objectbox.dart
@@ -4,3 +4,4 @@ export "src/common.dart";
 export "src/model.dart";
 export "src/store.dart";
 export "src/box.dart";
+export "src/bindings/constants.dart";