+++ lib/src/store.dart
@@ -1,35 +1,51 @@
 import "dart:ffi";
-import "dart:mirrors";
 
 import "bindings/bindings.dart";
 import "bindings/helpers.dart";
+import "ffi/cstring.dart";
 
 import "model.dart";
 
 class Store {
     Pointer<Void> _objectboxStore;
-    var _modelDescriptions;
+    Map<Type, Map<String, dynamic>> _modelDefinitions = {};
 
-    Store(var classes) {                        // TODO: allow setting options, e.g. database path
-        var model = Model(classes);
-        _modelDescriptions = model.desc;
+    Store(List<List<dynamic>> defs, {String directory, int maxDBSizeInKB, int fileMode, int maxReaders}) {                        // TODO: allow setting options, e.g. database path
+        defs.forEach((d) => _modelDefinitions[d[0]] = d[1]);
+        var model = Model(defs.map((d) => d[1]["model"] as Map<String, dynamic>).toList());
 
         var opt = bindings.obx_opt();
-        check(opt.address != 0);
+        checkObxPtr(opt, "failed to create store options");
         checkObx(bindings.obx_opt_model(opt, model.ptr));
+        if(directory != null && directory.length != 0) {
+            var cStr = new CString(directory);
+            checkObx(bindings.obx_opt_directory(opt, cStr.ptr));
+            cStr.free();
+        }
+        if(maxDBSizeInKB != null && maxDBSizeInKB > 0)
+            bindings.obx_opt_max_db_size_in_kb(opt, maxDBSizeInKB);
+        if(fileMode != null && fileMode >= 0)
+            bindings.obx_opt_file_mode(opt, fileMode);
+        if(maxReaders != null && maxReaders > 0)
+            bindings.obx_opt_max_readers(opt, maxReaders);
         _objectboxStore = bindings.obx_store_open(opt);
-        check(_objectboxStore != null);
-        check(_objectboxStore.address != 0);
+        checkObxPtr(_objectboxStore, "failed to create store");
     }
 
     close() {
         checkObx(bindings.obx_store_close(_objectboxStore));
     }
 
-    getEntityDescriptionFromClass(cls) {
-        final clsName = getSymbolName(reflectClass(cls).simpleName);
-        final idx = _modelDescriptions.indexWhere((e) => e["entity"]["name"] == clsName);
-        return idx == -1 ? null : _modelDescriptions[idx];
+    getEntityModelDefinitionFromClass(cls) {
+        return _modelDefinitions[cls]["model"];
+    }
+
+    getEntityReaderFromClass<T>() {
+        return _modelDefinitions[T]["reader"] as Map<String, dynamic> Function(T);
+    }
+
+    getEntityBuilderFromClass<T>() {
+        return _modelDefinitions[T]["builder"] as T Function(Map<String, dynamic>);
     }
 
     get ptr => _objectboxStore;