+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+0.2.0 (2019-09-11)
+------------------
+* UTF-8 support for Store and Box (thanks to [Buggaboo](https://github.com/Buggaboo) for [#14](https://github.com/objectbox/objectbox-dart/pull/14)!)
+* Bulk put and get functions (getMany, getAll, putMany)
+* Updated to objectbox-c 0.7
+* Basic Store options
+* Minimal unit tests
+* Removed reflection code, switched to model code generation instead
+* Minimal Flutter Desktop example for Dart 2.5.0
+
 0.1.0 (2019-09-03)
 ------------------
 * Minimal Store setup